pyproject.toml
[project]
name = "cleveland-parent-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "playwright>=1.58.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "schedule>=1.2.2",
]
